# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SignalGrid Device Policy & Security API",
    "description": "Enterprise device security, compliance, and policy enforcement platform. Integrates with MDM, NAC, ITSM, and SIEM systems.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@signalgrid.local"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://api.signalgrid.local",
      "description": "Production API",
      "variables": {}
    },
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/v1/devices": {
      "get": {
        "summary": "List devices",
        "operationId": "listDevices",
        "tags": ["Devices"],
        "security": [
          {
            "BearerAuth": []
          },
          {
            "APIKey": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "enrolled",
            "in": "query",
            "description": "Filter by enrollment status",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Device list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceList"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/devices/{deviceId}": {
      "get": {
        "summary": "Get device details",
        "operationId": "getDevice",
        "tags": ["Devices"],
        "security": [
          {
            "BearerAuth": []
          },
          {
            "APIKey": []
          }
        ],
        "parameters": [
          {
            "name": "deviceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Device details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Device"
                }
              }
            }
          },
          "404": {
            "description": "Device not found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from OIDC provider or admin API key"
      },
      "APIKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Admin API key for development (deprecated in production)"
      },
      "HMACSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Signature",
        "description": "HMAC-SHA256 request signature for device endpoints"
      }
    },
    "schemas": {
...
      "Device": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string"
          },
          "deviceSerial": {
            "type": "string"
          },
          "deviceModel": {
            "type": "string"
          },
          "osVersion": {
            "type": "string"
          },
          "enrolled": {
            "type": "boolean"
          },
          "enrolledAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastSeenAt": {
            "type": "string",
            "format": "date-time"
          },
          "managementId": {
            "type": "string"
          }
        }
      },
      "DeviceList": {
        "type": "object",
        "properties": {
          "devices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Device"
            }
          },
          "total": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          }
        }
      },
...
    },
...
  }
}